pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mbgen"
version = "1.0.0"
description = "MBGen: De Novo Molecular Generation from Mass Spectra via Many-Body Enhanced Diffusion"
requires-python = ">=3.9"
dependencies = [
    "hydra-core==1.3.2",
    "matplotlib==3.7.1",
    "numpy==1.23",
    "omegaconf==2.3.0",
    "overrides==7.3.1",
    "pandas==1.4",
    "pytorch_lightning==2.0.4",
    "setuptools==68.0.0",
    "torch_geometric==2.3.1",
    "torchmetrics==0.11.4",
    "tqdm",
    "tqdm_joblib",
    "wandb",
    "h5py",
    "seaborn",
    "myopic-mces"
]
authors = [
    {name = "Xichen Sun"},
    {name = "Wentao Wei"},
    {name = "Jiahua Rao"},
    {name = "Jiancong Xie"},
    {name = "Yuedong Yang"},
]
maintainers = [
    {name = "Xichen Sun"}
]
readme = "README.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["src"]
exclude = []
namespaces = false